refactor (CarFactory) : 용도에 맞는 변수명으로 변경
