+++ src/main/kotlin/racingcar/domain/RaceManager.kt
@@ -1,13 +1,14 @@
 package racingcar.domain
 
 import racingcar.racingcar.domain.CarFactory
+import racingcar.racingcar.domain.raceresult.RaceResult
 
 class RaceManager(
     private val numberGenerator: NumberGenerator,
 ) {
-    fun race(carFactory: CarFactory, raceCount: Int): CarFactory {
+    fun race(carFactory: CarFactory, raceCount: Int): RaceResult {
         repeat(raceCount) { carFactory.moveCars(numberGenerator) }
-        return carFactory
+        return RaceResult(carFactory.carNames, carFactory.record)
     }
 
     fun getWinners(carFactory: CarFactory): List<String> {
