+++ src/main/kotlin/racingcar/domain/raceresult/RaceResult.kt
@@ -1,0 +1,6 @@
+package racingcar.racingcar.domain.raceresult
+
+data class RaceResult(
+    val names: List<String>,
+    val result: List<List<Int>>,
+)
