+++ src/main/kotlin/racingcar/domain/raceresult/RaceResultDto.kt
@@ -1,4 +1,4 @@
-package racingcar.racingcar.domain
+package racingcar.racingcar.domain.raceresult
 
 data class RaceResultDto(
     val names: List<String>,
