+++ src/main/kotlin/racingcar/domain/raceresult/RaceResultMapper.kt
@@ -1,0 +1,10 @@
+package racingcar.racingcar.domain.raceresult
+
+class RaceResultMapper {
+    fun toRaceResultDto(raceResult: RaceResult): RaceResultDto {
+        return RaceResultDto(
+            names = raceResult.names,
+            result = raceResult.result,
+        )
+    }
+}
